Reunify lock and local repo transaction code and apply the same sorting
