+++ tests/Composer/Test/Fixtures/installer/abandoned-listed.test
@@ -27,8 +27,8 @@ update
 Loading composer repositories with package information
 Updating dependencies
 Lock file operations: 2 installs, 0 updates, 0 removals
-  - Installing c/c (1.0.0)
   - Installing a/a (1.0.0)
+  - Installing c/c (1.0.0)
 Writing lock file
 Installing dependencies from lock file (including require-dev)
 Package operations: 2 installs, 0 updates, 0 removals
