+++ tests/Composer/Test/Fixtures/installer/suggest-installed.test
@@ -22,8 +22,8 @@ update
 Loading composer repositories with package information
 Updating dependencies
 Lock file operations: 2 installs, 0 updates, 0 removals
-  - Installing b/b (1.0.0)
   - Installing a/a (1.0.0)
+  - Installing b/b (1.0.0)
 Writing lock file
 Installing dependencies from lock file (including require-dev)
 Package operations: 2 installs, 0 updates, 0 removals
