+++ tests/Composer/Test/Fixtures/installer/update-with-all-dependencies.test
@@ -50,8 +50,8 @@ update b/b --with-all-dependencies
 Loading composer repositories with package information
 Updating dependencies
 Lock file operations: 0 installs, 2 updates, 0 removals
-  - Updating b/b (1.0.0) to b/b (1.1.0)
   - Updating a/a (1.0.0) to a/a (1.1.0)
+  - Updating b/b (1.0.0) to b/b (1.1.0)
 Writing lock file
 Installing dependencies from lock file (including require-dev)
 Package operations: 0 installs, 2 updates, 0 removals
